Move late lint machanism in librustc_lint.
